requirements.txt
ollama==0.3.3
scrapy==2.11.2
scrapy-playwright==0.0.42
pyquery==2.0.1
pydantic==2.8.2
sentence-transformers==3.0.1
faiss-cpu==1.9.0
ollama==0.3.3
pysqlite3==0.5.4
pandas==2.2.3
uvicorn==0.30.1
websockets==14.1
